+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2022-07-20
+### Changed
+- `DatabaseResult.to_dataframe()` will now return an empty dataframe with the correct
+  column names if the result has column information but no rows. Warning: Pandas will
+  not be able to infer the correct dtypes from these columns.
+- Loosened version restrictions on `Pandas` and `ORJSON` dependencies to be any version number.
+
+
 ## [0.3.0] - 2022-01-21
 ### Added
 - Added support for multiple result sets, you can move between result sets using the new